+++ packages/contracts/src/__tests__/utils.spec.ts
@@ -1,7 +1,7 @@
 import { TypeDef } from '@dedot/codecs';
 import { describe, expect, it } from 'vitest';
 import { ContractTypeDef, ReturnFlags } from '../types/index.js';
-import { extractContractTypes, normalizeContractTypeDef, normalizeLabel, toReturnFlags } from '../utils.js';
+import { extractContractTypes, normalizeContractTypeDef, normalizeLabel, toReturnFlags } from '../utils/index.js';
 import { FLIPPER_CONTRACT_METADATA_V4 } from './contracts-metadata.js';
 
 describe('utils', () => {
