+++ packages/contracts/src/executor/abstract/DeployerExecutor.ts
@@ -4,7 +4,7 @@ import { GenericSubstrateApi } from '@dedot/types';
 import { HexString } from '@dedot/utils';
 import { TypinkRegistry } from '../../TypinkRegistry.js';
 import { ContractConstructorMessage, ExecutionOptions } from '../../types/index.js';
-import { normalizeLabel } from '../../utils.js';
+import { normalizeLabel } from '../../utils/index.js';
 import { Executor } from './Executor.js';
 
 export abstract class DeployerExecutor<ChainApi extends GenericSubstrateApi> extends Executor<ChainApi> {
