+++ packages/contracts/src/index.ts
@@ -1,4 +1,4 @@
-export * from './utils.js';
+export * from './utils/index.js';
 export * from './errors.js';
 export * from './executor/index.js';
 export * from './types/index.js';
